requirement.txt
streamlit
supabase
pillow
torch
torchvision
requests